+++ Packages/vChewing_MainAssembly/Package.swift
@@ -13,10 +13,11 @@ let package = Package(
     ),
   ],
   dependencies: [
+    .package(path: "../DanielGalasko_FolderMonitor"),
     .package(path: "../Jad_BookmarkManager"),
+    .package(path: "../Qwertyyb_ShiftKeyUpChecker"),
     .package(path: "../vChewing_CandidateWindow"),
     .package(path: "../vChewing_CocoaExtension"),
-    .package(path: "../DanielGalasko_FolderMonitor"),
     .package(path: "../vChewing_Hotenka"),
     .package(path: "../vChewing_IMKUtils"),
     .package(path: "../vChewing_LangModelAssembly"),
@@ -25,12 +26,11 @@ let package = Package(
     .package(path: "../vChewing_PhraseEditorUI"),
     .package(path: "../vChewing_PopupCompositionBuffer"),
     .package(path: "../vChewing_Shared"),
-    .package(path: "../Qwertyyb_ShiftKeyUpChecker"),
     .package(path: "../vChewing_SwiftExtension"),
     .package(path: "../vChewing_Tekkon"),
     .package(path: "../vChewing_TooltipUI"),
-    .package(path: "../vChewing_UpdateSputnik"),
     .package(path: "../vChewing_Uninstaller"),
+    .package(path: "../vChewing_UpdateSputnik"),
   ],
   targets: [
     .target(
@@ -39,8 +39,8 @@ let package = Package(
         .product(name: "BookmarkManager", package: "Jad_BookmarkManager"),
         .product(name: "CandidateWindow", package: "vChewing_CandidateWindow"),
         .product(name: "CocoaExtension", package: "vChewing_CocoaExtension"),
-        .product(name: "Hotenka", package: "vChewing_Hotenka"),
         .product(name: "FolderMonitor", package: "DanielGalasko_FolderMonitor"),
+        .product(name: "Hotenka", package: "vChewing_Hotenka"),
         .product(name: "IMKUtils", package: "vChewing_IMKUtils"),
         .product(name: "LangModelAssembly", package: "vChewing_LangModelAssembly"),
         .product(name: "Megrez", package: "vChewing_Megrez"),
@@ -52,8 +52,8 @@ let package = Package(
         .product(name: "SwiftExtension", package: "vChewing_SwiftExtension"),
         .product(name: "Tekkon", package: "vChewing_Tekkon"),
         .product(name: "TooltipUI", package: "vChewing_TooltipUI"),
-        .product(name: "UpdateSputnik", package: "vChewing_UpdateSputnik"),
         .product(name: "Uninstaller", package: "vChewing_Uninstaller"),
+        .product(name: "UpdateSputnik", package: "vChewing_UpdateSputnik"),
       ]
     ),
     .testTarget(
