+++ Packages/vChewing_MainAssembly/README.md
@@ -1,3 +1,3 @@
 # MainAssembly
 
-威注音輸入法核心元件。
+威注音輸入法主體總成模組。
