--- Packages/vChewing_MainAssembly/Sources/MainAssembly/MainAssembly.swift
@@ -1,5 +1,0 @@
-public struct MainAssembly {
-  public private(set) var text = "Hello, World!"
-
-  public init() {}
-}
