+++ Packages/vChewing_MainAssembly/Tests/MainAssemblyTests/MainAssemblyTests.swift
@@ -3,9 +3,5 @@ import XCTest
 
 final class MainAssemblyTests: XCTestCase {
   func testExample() throws {
-    // This is an example of a functional test case.
-    // Use XCTAssert and related functions to verify your tests produce the correct
-    // results.
-    XCTAssertEqual(MainAssembly().text, "Hello, World!")
   }
 }
