+++ Source/Modules/AppDelegateImpl.swift
@@ -1,0 +1,20 @@
+// (c) 2011 and onwards The OpenVanilla Project (MIT License).
+// All possible vChewing-specific modifications are of:
+// (c) 2021 and onwards The vChewing Project (MIT-NTL License).
+// ====================
+// This code is released under the MIT license (SPDX-License-Identifier: MIT)
+// ... with NTL restriction stating that:
+// No trademark license is granted to use the trade names, trademarks, service
+// marks, or product names of Contributor, except as required to fulfill notice
+// requirements defined in MIT License.
+
+import AppKit
+import MainAssembly
+
+extension AppDelegate {
+  // New About Window
+  @IBAction func about(_: Any) {
+    CtlAboutWindow.show()
+    NSApp.popup()
+  }
+}
