+++ Source/Modules/UIModules/PrefUI/CtlPrefUIShared.swift
@@ -6,6 +6,7 @@
 // marks, or product names of Contributor, except as required to fulfill notice
 // requirements defined in MIT License.
 
+import MainAssembly
 import SSPreferences
 import SwiftExtension
 import SwiftUI
