+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneGeneral.swift
@@ -6,6 +6,7 @@
 // marks, or product names of Contributor, except as required to fulfill notice
 // requirements defined in MIT License.
 
+import MainAssembly
 import Shared
 import SSPreferences
 import SwiftExtension
