+++ Source/Modules/UIModules/PrefUITabs.swift
@@ -7,6 +7,7 @@
 // requirements defined in MIT License.
 
 import AppKit
+import MainAssembly
 import Shared
 
 enum PrefUITabs: String, CaseIterable {
