+++ Source/Modules/WindowControllers/CtlPrefWindow_PhraseEditor.swift
@@ -9,6 +9,7 @@
 import AppKit
 import Foundation
 import LangModelAssembly
+import MainAssembly
 import PhraseEditorUI
 import Shared
